# backend/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Express",
    "description": "Documentation for the backoffice project"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Allows employees to sign up.",
        "responses": {
          "302": {
            "description": "Successful login. Redirect to /employees"
          },
          "401": {
            "description": "Invalid credentials. Redirect to /login"
          },
          "500": {
            "description": "Internal Server Error. Rendering error page"
          }
        }
      },
      "parameters": [
        {
          "in": "body",
          "name": "User",
          "schema": {
            "$ref": "#/definitions/User"
          }
        }
      ]
    },
...
    "/employees": {
      "get": {
        "security": [
          {
            "in": "cookie",
            "name": "access_token"
          }
        ],
        "tags": [
          "Employees"
        ],
        "summary": "Return a list of all employees",
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Employee"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/places/all": {
      "security": [
        {
          "in": "cookie",
          "name": "access_token"
        }
      ],
      "get": {
        "tags": [
          "Locals"
        ],
        "summary": "Get all local data",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Page number (optional)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
},
"definitions": {
"Employee": {
"required": [
"name",
"role",
"email",
"password"
],
"properties": {
"name": {
"type": "string",
"description": "Employee's name"
},
"role": {
"type": "string",
"description": "Employee's role. Must be either 'admin' or 'employee'",
"enum": [
"admin",
"employee"
]
},
"email": {
"type": "string",
"description": "Employee's email"
},
"password": {
"type": "string",
"description": "Employee's password. Must contain at least 8 characters, one uppercase letter, one lowercase letter, and one number"
},
"deleted": {
"type": "boolean",
"description": "Indicates if the employee is active or deleted. 'true' means deleted, 'false' means active."
}
}
},
"Event": {
"required": [
"title",
"description",
"location",
"startTime",
"endTime",
"eventType",
"image"
],
"properties": {
"title": {
"type": "string",
"description": "The title of the event"
},
"description": {
"type": "string",
"description": "The description of the event"
},
"location": {
"type": "string",
"description": "The location of the event"
},
"startTime": {
"type": "string",
"format": "date-time",
"description": "The start time of the event"
},
"endTime": {
"type": "string",
"format": "date-time",
"description": "The end time of the event"
},
"eventType": {
"type": "string",
"enum": [
"free",
"paid"
],
"default": "free",
"description": "The type of the event. Can be 'free' or 'paid'"
},
"price": {
"type": "number",
"default": 0,
"minimum": 0,
"description": "The price of the event. Only applicable for 'paid' events"
},
"image": {
"type": "string",
"description": "The image URL of the event"
},
"deleted": {
"type": "boolean",
"default": false,
"description": "Indicates if the event is active or deleted. 'true' means deleted, 'false' means active"
}
}
},
"Local": {
"required": [
"localdata"
],
"properties": {
"localdata": {
"type": "object",
"description": "The data of the local"
},
"eventsOcurring": {
"type": "array",
"items": {
"$ref": "#/definitions/Event"
},
"description": "The events occurring at the local"
},
"price": {
"type": "number",
"description": "The price of the local"
},
"deleted": {
"type": "boolean",
"default": false,
"description": "Indicates if the local is active or deleted. 'true' means deleted, 'false' means active"
}
}
},
"Ticket": {
"required": [
"employeeID",
"date",
"free",
"users"
],
"properties": {
"events": {
"type": "string",
"description": "The ID of the event associated with the ticket",
"$ref": "#/definitions/Event"
},
"locals": {
"type": "string",
"description": "The ID of the local associated with the ticket",
"$ref": "#/definitions/Local"
},
"employeeID": {
"type": "string",
"description": "The ID of the employee associated with the ticket",
"format": "uuid",
"$ref": "#/definitions/Employee"
},
"price": {
"type": "number",
"description": "The price of the ticket"
},
"date": {
"type": "string",
"format": "date",
"description": "The date of the ticket"
},
"free": {
"type": "boolean",
"description": "Indicates if the ticket is free or not"
},
"users": {
"type": "array",
"items": {
"$ref": "#/definitions/User"
},
"description": "The IDs of the users associated with the ticket"
}
}
},
"User": {
"required": [
"username",
"password",
"email",
"firstName",
"lastName",
"dateOfBirth"
],
"properties": {
"username": {
"type": "string",
"description": "Username of the user"
},
"password": {
"type": "string",
"description": "Password of the user (must contain at least 8 characters, including uppercase, lowercase, and a number)"
},
"email": {
"type": "string",
"description": "Email address of the user"
},
"firstName": {
"type": "string",
"description": "First name of the user"
},
"lastName": {
"type": "string",
"description": "Last name of the user"
},
"dateOfBirth": {
"type": "string",
"description": "Date of birth of the user"
},
"createdAt": {
"type": "string",
"description": "Date and time when the user was created"
},
"tickets": {
"type": "array",
"items": {
"$ref": "#/definitions/Ticket"
},
"description": "List of tickets associated with the user"
},
"balance": {
"type": "number",
"description": "Balance amount of the user"
},
"points": {
"type": "number",
"description": "Points earned by the user"
},
"deleted": {
"type": "boolean",
"description": "Indicates if the user is deleted or not"
}
}
}
}
}